separate blueprints for api vs mai
