+++ app/blueprints/api/__init__.py
@@ -1,0 +1,12 @@
+from apiflask import APIBlueprint
+
+api = APIBlueprint("api", __name__)
+
+from . import (  # noqa: E402, F401
+    harvest_jobs,
+    harvest_records,
+    harvest_sources,
+    organizations,
+    query,
+    validate,
+)
