+++ app/routes.py
@@ -1,2033 +1,19 @@
-import csv
-import json
-import logging
-import os
-import secrets
-import tempfile
-import time
-import uuid
-from datetime import timedelta
-from functools import wraps
+"""Backward-compatible re-exports; route registration lives in app.blueprints."""
 
-import jwt
-import requests
-from apiflask import APIBlueprint
-from apiflask.schemas import FileSchema
-from cryptography.hazmat.backends import default_backend
-from cryptography.hazmat.primitives.serialization import load_pem_private_key
-from dotenv import load_dotenv
-from flask import (
-    Blueprint,
-    Response,
-    current_app,
-    flash,
-    g,
-    jsonify,
-    make_response,
-    redirect,
-    render_template,
-    request,
-    session,
-    url_for,
+from app.blueprints import register_routes
+from app.blueprints.deps import (
+    UnsafeTemplateEnvError,
+    db,
+    load_manager,
+    render_block,
 )
-from markupsafe import escape
-from marshmallow import ValidationError
-
-from database.interface import HarvesterDBInterface
-from harvester.lib.load_manager import LoadManager
-from harvester.utils.general_utils import (
-    convert_to_int,
-    dynamic_map_list_items_to_dict,
-    get_datetime,
-    is_it_true,
-    is_valid_uuid4,
-    process_job_complete_percentage,
-)
-
-from . import current_unix_timestamp, htmx
-from .api_schemas import (
-    ErrorInfo,
-    JobInfo,
-    OrgCreate,
-    OrgInfo,
-    QueryInfo,
-    RecordInfo,
-    SourceInfo,
-    ValidationErrorResponseSchema,
-    ValidationResultSchema,
-    ValidatorInfo,
-)
-from .auth import LoginRequiredAuth
-from .forms import (
-    DatasetSlugForm,
-    HarvestSourceForm,
-    HarvestTriggerForm,
-    LocalDevLoginForm,
-    OrganizationForm,
-    OrganizationTriggerForm,
-    ValidatorForm,
-)
-from .local_dev_auth import (
-    LOCAL_DEV_SESSION_EMAIL,
-    is_local_dev_login_enabled,
-    validate_local_dev_credentials,
-)
-from .paginate import Pagination
-from .util import (
-    fetch_json_from_url,
-    make_new_org_contract,
-    make_new_record_error_contract,
-    make_new_source_contract,
-    validate_records,
-)
-
-logger = logging.getLogger("harvest_admin")
-
-main = Blueprint("main", __name__)
-api = APIBlueprint("api", __name__)
-auth = LoginRequiredAuth()
-login_required = auth.login_required()
-
-
-db = HarvesterDBInterface()
-
-load_manager = LoadManager()
-
-# Login authentication
-load_dotenv()
-CKAN_URL = os.getenv("CKAN_URL") or ""
-CLIENT_ID = os.getenv("CLIENT_ID")
-REDIRECT_URI = os.getenv("REDIRECT_URI")
-ISSUER = os.getenv("ISSUER")
-AUTH_URL = ISSUER + "/openid_connect/authorize"
-TOKEN_URL = ISSUER + "/api/openid_connect/token"
-
-
-STATUS_STRINGS_ENUM = {404: "Not Found"}
-
-
-def JSON_NOT_FOUND():
-    """Return our most generic error response.
-
-    This is a function rather than a constant because `jsonify` requires
-    an app context.
-    """
-    response = jsonify({"error": STATUS_STRINGS_ENUM[404]})
-    response.status_code = 404
-    return response
-
-
-class UnsafeTemplateEnvError(RuntimeError):
-    pass
-
-
-def render_block(template_name: str, block_name: str, **context) -> Response:
-    """
-    Render a specific block from a Jinja template, while using the Flask's default environment.
-    """
-    env = current_app.jinja_env
-    if not getattr(env, "autoescape", None):
-        raise UnsafeTemplateEnvError(
-            "Jinja autoescape is disabled; enable it or use Flask's jinja_env."
-        )
-    template = env.get_template(template_name)
-
-    # Render only the named block (Jinja will still escape vars inside the block)
-    block_gen = template.blocks[block_name]
-    html = "".join(block_gen(template.new_context(context)))
-    return Response(html, mimetype="text/html; charset=utf-8")
-
-
-def valid_id_required(f):
-    """Decorator to check that ALL function arguments are valid IDs.
-
-    Returns a JSON 404 if they are not.
-
-    TODO: Make the decorator return a pretty templated HTML page if the user
-    would be expecting a 404 page.
-    """
-
-    @wraps(f)
-    def decorated_function(*args, **kwargs):
-        for arg in args:
-            if not is_valid_uuid4(arg):
-                return JSON_NOT_FOUND()
-        for kwarg in kwargs.values():
-            if not is_valid_uuid4(kwarg):
-                return JSON_NOT_FOUND()
-        return f(*args, **kwargs)
-
-    return decorated_function
-
-
-def _get_org_by_identifier(org_identifier: str):
-    """Resolve an organization by UUID id or slug."""
-    org = None
-    if is_valid_uuid4(org_identifier):
-        org = db.get_organization(org_identifier)
-    if org is None:
-        org = db.get_organization_by_slug(org_identifier)
-    return org
-
-
-def _get_org_url_identifier(org) -> str | None:
-    """Prefer an organization's slug for URLs, with UUID as a fallback."""
-    if org is None:
-        return None
-    return org.slug or org.id
-
-
-def _get_request_actor() -> str:
-    actor = getattr(g, "request_actor", None) or session.get("user")
-    return actor or "<anonymous>"
-
-
-def _get_request_auth_type() -> str:
-    auth_type = getattr(g, "request_auth_type", None)
-    if auth_type:
-        return auth_type
-    if session.get("user"):
-        return "session"
-    return "anonymous"
-
-
-def _log_mutation(action: str, entity: str, entity_id: str | None = None, **details):
-    detail_fields = {
-        "user": _get_request_actor(),
-        "auth_type": _get_request_auth_type(),
-        "method": request.method,
-        "path": request.path,
-    }
-    if entity_id is not None:
-        detail_fields[f"{entity}_id"] = entity_id
-    detail_fields.update(details)
-    ordered_keys = sorted(detail_fields)
-    message = "Audit %s %s " % (action, entity) + " ".join(
-        f"{key}=%s" for key in ordered_keys
-    )
-    logger.info(message, *(detail_fields[key] for key in ordered_keys))
-
-
-def create_client_assertion():
-    private_key_data = os.getenv("OPENID_PRIVATE_KEY")
-    if not private_key_data:
-        raise ValueError("No private key found in the environment variable")
-
-    private_key = load_pem_private_key(
-        private_key_data.encode("utf-8"), password=None, backend=default_backend()
-    )
-
-    now = int(time.time())
-    payload = {
-        "iss": CLIENT_ID,
-        "sub": CLIENT_ID,
-        "aud": TOKEN_URL,
-        "jti": uuid.uuid4().hex,
-        "exp": now + 900,  # Token is valid for 15 minutes
-        "iat": now - 100,  # Issued at time, allow some leeway
-    }
-
-    return jwt.encode(payload, private_key, algorithm="RS256")
-
-
-# HELPERS
-# TODO: when refactoring routes put them in their respective files
-## HARVEST SOURCE HELPERS
-def trigger_manual_job_helper(source_id, job_type="harvest"):
-    message = load_manager.trigger_manual_job(source_id, job_type)
-    flash(message)
-    return redirect(url_for("api.view_harvest_source", source_id=source_id))
-
-
-def _redirect_to_login_gov():
-    state = secrets.token_urlsafe(32)
-    nonce = secrets.token_urlsafe(32)
-    session["state"] = state
-    session["nonce"] = nonce
-    logger.info(
-        "Login initiated from ip=%s next=%s",
-        request.headers.get("X-Forwarded-For", request.remote_addr),
-        session.get("next"),
-    )
-
-    auth_request_url = (
-        f"{AUTH_URL}?response_type=code"
-        f"&acr_values=urn:acr.login.gov:auth-only+http://idmanagement.gov/ns/assurance/aal/2?hspd12=true"
-        f"&client_id={CLIENT_ID}"
-        f"&nonce={nonce}"
-        "&prompt=select_account"
-        f"&redirect_uri={REDIRECT_URI}"
-        f"&scope=openid+email"
-        f"&state={state}"
-    )
-    return redirect(auth_request_url)
-
-
-@main.route("/login/oidc", methods=["GET"])
-def login_oidc():
-    return _redirect_to_login_gov()
-
-
-@main.route("/login", methods=["GET", "POST"])
-def login():
-    if not is_local_dev_login_enabled():
-        if request.method == "POST":
-            return make_response("Not Found", 404)
-        return _redirect_to_login_gov()
-
-    form = LocalDevLoginForm(meta={"csrf": False})
-    if request.method == "POST":
-        username = request.form.get("username", "")
-        password = request.form.get("password", "")
-        if validate_local_dev_credentials(username, password):
-            session["user"] = LOCAL_DEV_SESSION_EMAIL
-            session["last_activity"] = current_unix_timestamp()
-            logger.info(
-                "Local dev login succeeded for user=%s ip=%s",
-                LOCAL_DEV_SESSION_EMAIL,
-                request.headers.get("X-Forwarded-For", request.remote_addr),
-            )
-            next_url = session.pop("next", None)
-            if next_url:
-                return redirect(next_url)
-            return redirect(url_for("main.index"))
-        flash("Invalid username or password.", "danger")
-        logger.warning(
-            "Local dev login failed from ip=%s",
-            request.headers.get("X-Forwarded-For", request.remote_addr),
-        )
-
-    return render_template("local_login.html", form=form)
-
-
-@main.route("/logout")
-def logout():
-    user = session.get("user")
-    session.clear()
-    logger.info(
-        "Logout completed for user=%s ip=%s",
-        user or "<anonymous>",
-        request.headers.get("X-Forwarded-For", request.remote_addr),
-    )
-    return redirect(url_for("main.index"))
-
-
-@main.route("/callback")
-def callback():
-    code = request.args.get("code")
-    state = request.args.get("state")
-    request_ip = request.headers.get("X-Forwarded-For", request.remote_addr)
-
-    logger.info(
-        "Login callback received ip=%s has_code=%s has_state=%s",
-        request_ip,
-        bool(code),
-        bool(state),
-    )
-
-    if state != session.pop("state", None):
-        logger.warning("Login callback failed: state mismatch ip=%s", request_ip)
-        return "State mismatch error", 400
-
-    try:
-        client_assertion = create_client_assertion()
-    except Exception as e:
-        logger.exception(
-            "Login callback failed: could not create client assertion ip=%s error=%s",
-            request_ip,
-            repr(e),
-        )
-        return "Failed to prepare login request", 500
-
-    # ruff: noqa: E501
-    token_payload = {
-        "grant_type": "authorization_code",
-        "code": code,
-        "redirect_uri": REDIRECT_URI,
-        "client_id": CLIENT_ID,
-        "client_assertion_type": "urn:ietf:params:oauth:client-assertion-type:jwt-bearer",
-        "client_assertion": client_assertion,
-    }
-    try:
-        response = requests.post(TOKEN_URL, data=token_payload)
-    except Exception as e:
-        logger.exception(
-            "Login callback failed: token exchange request error ip=%s error=%s",
-            request_ip,
-            repr(e),
-        )
-        return "Failed to fetch access token", 400
-
-    if response.status_code != 200:
-        logger.error(
-            "Login callback failed: token endpoint returned status=%s ip=%s body=%s",
-            response.status_code,
-            request_ip,
-            response.text,
-        )
-        return "Failed to fetch access token", 400
-
-    try:
-        token_data = response.json()
-    except ValueError as e:
-        logger.exception(
-            "Login callback failed: token endpoint returned invalid JSON ip=%s error=%s",
-            request_ip,
-            repr(e),
-        )
-        return "Failed to fetch access token", 400
-
-    id_token = token_data.get("id_token")
-    if not id_token:
-        logger.error(
-            "Login callback failed: id_token missing from token response ip=%s",
-            request_ip,
-        )
-        return "Failed to fetch access token", 400
-
-    try:
-        decoded_id_token = jwt.decode(id_token, options={"verify_signature": False})
-    except Exception as e:
-        logger.exception(
-            "Login callback failed: unable to decode id_token ip=%s error=%s",
-            request_ip,
-            repr(e),
-        )
-        return "Failed to decode login token", 400
-
-    usr_email = decoded_id_token.get("email")
-    if not usr_email:
-        logger.error(
-            "Login callback failed: email missing from id_token claims ip=%s sub=%s",
-            request_ip,
-            decoded_id_token.get("sub"),
-        )
-        return "Failed to identify user", 400
-
-    usr_email = usr_email.lower()
-    usr_ssoid = decoded_id_token.get("sub")
-    if not usr_ssoid:
-        logger.error(
-            "Login callback failed: sub missing from id_token claims user=%s ip=%s",
-            usr_email,
-            request_ip,
-        )
-        return "Failed to identify user", 400
-
-    usr_info = {"email": usr_email, "ssoid": usr_ssoid}
-    usr = db.verify_user(usr_info)
-
-    if usr:
-        session["user"] = usr_email
-        session["last_activity"] = current_unix_timestamp()
-        logger.info(
-            "Login succeeded for user=%s ssoid=%s ip=%s",
-            usr_email,
-            usr_ssoid,
-            request_ip,
-        )
-        next_url = session.pop("next", None)
-        if next_url:
-            return redirect(url_for(next_url))
-        else:
-            return redirect(url_for("main.index"))
-    else:
-        logger.warning(
-            "Login rejected for unregistered user=%s ssoid=%s ip=%s",
-            usr_email,
-            usr_ssoid,
-            request_ip,
-        )
-        flash("Please request registration from the admin before proceeding.")
-        return redirect(url_for("main.index"))
-
-
-# Routes
-@main.route("/", methods=["GET"])
-def index():
-    return redirect(url_for("main.organization_list"))
-
-
-## Organizations
-@main.route("/organization_list/", methods=["GET"])
-def organization_list():
-    organizations = db.get_all_organizations()
-    data = {"organizations": organizations}
-    return render_template("view_org_list.html", data=data)
-
-
-### Get Org List (JSON)
-@api.get("/api/organization_list/")
-@api.doc(
-    responses={
-        200: {
-            "description": "List organizations",
-            "content": {
-                "application/json": {"schema": {"type": "array", "items": OrgInfo}}
-            },
-        }
-    }
-)
-def organization_list_api():
-    """Return the JSON list of all organizations.
-
-    The HTML list page lives on the main blueprint at
-    `GET /organization_list/` (see `organization_list`).
-    """
-    return jsonify(db._to_dict(db.get_all_organizations()))
-
-
-### Add Org (JSON)
-@api.post("/api/organization/add")
-@api.doc(hide=True)  # don't include the authenticated API
-@api.input(OrgCreate, validation=False)  # validated explicitly for API POST
-@login_required
-def add_organization_api(**kwargs):
-    org_data = request.json or {}
-    try:
-        org_data = OrgCreate().load(org_data)
-    except ValidationError as e:
-        return make_response(jsonify({"detail": e.messages}), 422)
-
-    org = db.add_organization(org_data)
-    if org:
-        _log_mutation("create", "organization", org.id, organization_slug=org.slug)
-        return make_response(
-            jsonify({"message": f"Added new organization with ID: {org.id}"}), 200
-        )
-    else:
-        return make_response(jsonify({"error": "Failed to add organization."}), 400)
-
-
-### Add Org (HTML form)
-@main.route("/organization/add", methods=["GET", "POST"])
-@login_required
-def add_organization():
-    form = OrganizationForm(db_interface=db)
-    if form.validate_on_submit():
-        new_org = make_new_org_contract(form)
-        org = db.add_organization(new_org)
-        if org:
-            _log_mutation("create", "organization", org.id, organization_slug=org.slug)
-            flash(f"Added new organization with ID: {org.id}")
-        else:
-            flash("Failed to add organization.")
-        return redirect(url_for("main.organization_list"))
-    elif form.errors:
-        flash(form.errors)
-        return redirect(url_for("main.add_organization"))
-    return render_template(
-        "edit_data.html",
-        form=form,
-        action="Add",
-        data_type="Organization",
-        button="Submit",
-    )
-
-
-### View Org (HTML detail page)
-@main.get("/organization/<string:org_identifier>")
-def view_organization(org_identifier: str):
-    """Render the HTML organization detail page by UUID or slug.
-
-    This route used to accept only UUIDs. It now resolves either a UUID or a
-    slug to an organization, then prefers the slug when linking back to the
-    organization page.
-
-    The JSON representation lives on the API blueprint at
-    `GET /api/organization/<org_identifier>` (see `get_organization`).
-    """
-    org = _get_org_by_identifier(org_identifier)
-    org_id = org.id if org is not None else org_identifier
-
-    # Only authenticated users get the edit/delete action form.
-    form = OrganizationTriggerForm() if session.get("user") else None
-    sources = db.get_harvest_source_by_org(org_id)
-    future_harvest_jobs = {}
-    for source in sources:
-        job = db.get_new_harvest_jobs_by_source_in_future(source.id)
-        if len(job):
-            future_harvest_jobs[source.id] = job[0].date_created
-    harvest_jobs = {}
-    for source in sources:
-        job = db.get_first_harvest_job_by_filter(
-            {"harvest_source_id": source.id, "status": "complete"}
-        )
-        if job:
-            harvest_jobs[source.id] = job
-    data = {
-        "ckan_url": CKAN_URL,
-        "organization": org,
-        "organization_dict": db._to_dict(org),
-        "harvest_sources": sources,
-        "harvest_jobs": harvest_jobs,
-        "future_harvest_jobs": future_harvest_jobs,
-    }
-    return render_template(
-        "view_org_data.html",
-        data=data,
-        form=form,
-    ), (200 if org is not None else 404)
-
-
-@main.post("/organization/<string:org_identifier>")
-@login_required
-def update_organization_actions(org_identifier: str):
-    """Handle authenticated organization action form submissions."""
-    org = _get_org_by_identifier(org_identifier)
-    org_id = org.id if org is not None else org_identifier
-    org_url_identifier = _get_org_url_identifier(org) or org_identifier
-    form = OrganizationTriggerForm()
-
-    if not form.validate_on_submit():
-        flash(form.errors)
-        return redirect(
-            url_for("main.view_organization", org_identifier=org_url_identifier)
-        )
-
-    if form.edit.data:
-        return redirect(url_for("api.edit_organization", org_id=org_id))
-    elif form.delete.data:
-        try:
-            message, status = db.delete_organization(org_id)
-            _log_mutation(
-                "delete",
-                "organization",
-                org_id,
-                organization_slug=org_url_identifier,
-                status=status,
-            )
-            flash(message)
-            if status == 409:
-                return redirect(
-                    url_for(
-                        "main.view_organization",
-                        org_identifier=org_url_identifier,
-                    )
-                )
-            else:
-                return redirect(url_for("main.organization_list"))
-        except Exception as e:
-            message = f"Failed to delete organization :: {repr(e)}"
-            logger.error(message)
-            flash(message)
-            return redirect(
-                url_for(
-                    "main.view_organization",
-                    org_identifier=org_url_identifier,
-                )
-            )
-
-    return redirect(
-        url_for("main.view_organization", org_identifier=org_url_identifier)
-    )
-
-
-### Edit Org (JSON)
-@api.post("/api/organization/edit/<string:org_id>")
-@api.doc(hide=True)  # don't document the authenticated API
-@login_required
-@valid_id_required
-def edit_organization_api(org_id):
-    org = db.update_organization(org_id, request.json)
-    if org:
-        _log_mutation("edit", "organization", org.id, organization_slug=org.slug)
-        return {"message": f"Updated org with ID: {org.id}"}, 200
-    else:
-        return {"error": "Failed to update organization."}, 400
-
-
-### Edit Org (HTML form)
-@api.route("/organization/edit/<string:org_id>", methods=["GET", "POST"])
-@api.doc(hide=True)  # don't document the authenticated API
-@login_required
-@valid_id_required  # TODO: Use an HTML 404 page
-def edit_organization(org_id):
-    org = db._to_dict(db.get_organization(org_id))
-    form = OrganizationForm(organization_id=org_id, data=org, db_interface=db)
-    if form.validate_on_submit():
-        new_org_data = make_new_org_contract(form)
-        org = db.update_organization(org_id, new_org_data)
-        if org:
-            _log_mutation("edit", "organization", org.id, organization_slug=org.slug)
-            flash(f"Updated org with ID: {org.id}")
-        else:
-            flash("Failed to update organization.")
-        return redirect(
-            url_for(
-                "main.view_organization",
-                org_identifier=_get_org_url_identifier(org) or org_id,
-            )
-        )
-    elif form.errors:
-        flash(form.errors)
-        return redirect(url_for("api.edit_organization", org_id=org_id))
-
-    return render_template(
-        "edit_data.html",
-        form=form,
-        action="Edit",
-        data_type="Organization",
-        button="Update",
-    )
-
-
-### Get Org (JSON detail)
-@api.get("/api/organization/<string:org_identifier>")
-@api.doc(
-    responses={
-        200: {
-            "description": "View organization info",
-            "content": {"application/json": {"schema": OrgInfo}},
-        }
-    }
-)
-def get_organization(org_identifier: str):
-    """Return the JSON representation of an organization by UUID or slug.
-
-    The HTML detail page lives on the main blueprint at
-    `GET /organization/<org_identifier>` (see `view_organization`).
-    """
-    org = _get_org_by_identifier(org_identifier)
-    if org is None:
-        return make_response(jsonify({"message": "Organization not found"}), 404)
-    return jsonify(org.to_dict())
-
-
-### Delete Org
-@api.route("/api/organization/<string:org_id>", methods=["DELETE"])
-@api.doc(hide=True)  # don't include the authenticated API
-@login_required
-@valid_id_required
-def delete_organization(org_id):
-    try:
-        message, status = db.delete_organization(org_id)
-        _log_mutation("delete", "organization", org_id, status=status)
-        return make_response(jsonify({"message": message}), status)
-    except Exception as e:
-        message = f"Failed to delete organization :: {repr(e)}"
-        logger.error(message)
-        return make_response(jsonify({"message": message}), 500)
-
-
-## Harvest Source
-### Add Source (JSON)
-@api.post("/api/harvest_source/add")
-@api.doc(hide=True)  # don't document the authenticated API
-@login_required
-def add_harvest_source_api():
-    try:
-        source = db.add_harvest_source(request.json)
-        job_message = load_manager.schedule_first_job(source.id)
-        if source and job_message:
-            _log_mutation(
-                "create",
-                "harvest_source",
-                source.id,
-                organization_id=source.organization_id,
-                source_name=source.name,
-            )
-            return make_response(
-                jsonify(
-                    {
-                        "message": f"Added new harvest source with ID: {source.id}. {job_message}"
-                    }
-                ),
-                200,
-            )
-    except Exception as e:
-        message = "Failed to add harvest source."
-        logger.error(f"{message} :: {repr(e)}")
-        return make_response(jsonify({"message": message}), 500)
-
-
-### Add Source (HTML form)
-@api.route("/harvest_source/add", methods=["POST", "GET"])
-@api.doc(hide=True)  # don't document the authenticated API
-@login_required
-def add_harvest_source():
-    form = HarvestSourceForm()
-    organizations = db.get_all_organizations()
-    organization_choices = sorted(
-        [(str(org.id), f"{org.name} - {org.id}") for org in organizations],
-        key=lambda x: x[1].lower(),
-    )
-    form.organization_id.choices = organization_choices
-    if form.validate_on_submit():
-        new_source = make_new_source_contract(form)
-        source = db.add_harvest_source(new_source)
-        job_message = load_manager.schedule_first_job(source.id)
-        if source and job_message:
-            _log_mutation(
-                "create",
-                "harvest_source",
-                source.id,
-                organization_id=source.organization_id,
-                source_name=source.name,
-            )
-            flash(f"Added new harvest source with ID: {source.id}. {job_message}")
-        else:
-            flash("Failed to add harvest source.")
-        return redirect(url_for("main.harvest_source_list"))
-    elif form.errors:
-        flash(form.errors)
-        return redirect(url_for("api.add_harvest_source"))
-    return render_template(
-        "edit_data.html",
-        form=form,
-        action="Add",
-        data_type="Harvest Source",
-        button="Submit",
-    )
-
-
-@api.get("/harvest_source/<source_id>")
-@api.doc(
-    responses={  # HTML or JSON output so specify manual response
-        200: {
-            "description": "View harvest source",
-            "content": {"application/json": {"schema": SourceInfo}},
-        }
-    }
-)
-@valid_id_required  # TODO: Use an HTML 404 page
-def view_harvest_source(source_id: str):
-    jobs_htmx_vars = {
-        "target_div": "#paginated__harvest-jobs",
-        "endpoint_url": f"/harvest_source/{source_id}",
-        "page_param": "page",
-    }
-    datasets_htmx_vars = {
-        "target_div": "#paginated__datasets",
-        "endpoint_url": f"/harvest_source/{source_id}",
-        "page_param": "datasets_page",
-    }
-    harvest_jobs_facets = (
-        f"harvest_source_id eq {source_id},date_created le {get_datetime()}"
-    )
-    jobs_count = db.pget_harvest_jobs(
-        facets=harvest_jobs_facets,
-        count=True,
-    )
-
-    pagination = Pagination(
-        count=jobs_count,
-        current=request.args.get("page", 1, type=convert_to_int),
-    )
-
-    jobs = db.pget_harvest_jobs(
-        facets=harvest_jobs_facets,
-        page=pagination.db_current,
-        order_by="desc",
-    )
-
-    if htmx:
-        htmx_target = request.headers.get("HX-Target", "")
-
-        if "paginated__datasets" in htmx_target:
-            datasets_pagination = Pagination(
-                count=db.get_datasets_by_source(source_id=source_id, count=True),
-                current=request.args.get("datasets_page", 1, type=convert_to_int),
-            )
-            datasets = db.get_datasets_by_source(
-                source_id=source_id,
-                page=datasets_pagination.db_current,
-                per_page=datasets_pagination.per_page,
-            )
-            data = {
-                "source": {"id": source_id},
-                "datasets": datasets,
-                "datasets_htmx_vars": datasets_htmx_vars,
-            }
-            logger.info(
-                "Rendered harvest source datasets partial source_id=%s page=%s "
-                "datasets_count=%s returned=%s",
-                source_id,
-                datasets_pagination.current,
-                datasets_pagination.count,
-                len(datasets),
-            )
-            return render_block(
-                "view_source_data.html",
-                "htmx_paginated_datasets",
-                data=data,
-                pagination=datasets_pagination.to_dict(),
-                datasets_pagination=datasets_pagination.to_dict(),
-            )
-        else:
-            data = {
-                "source": {"id": source_id},
-                "jobs": jobs,
-                "htmx_vars": jobs_htmx_vars,
-            }
-            logger.info(
-                "Rendered harvest source jobs partial source_id=%s page=%s jobs_count=%s "
-                "returned=%s",
-                source_id,
-                pagination.current,
-                pagination.count,
-                len(jobs),
-            )
-            return render_block(
-                "view_source_data.html",
-                "htmx_paginated",
-                data=data,
-                pagination=pagination.to_dict(),
-            )
-    else:
-        form = HarvestTriggerForm()
-        records_count = db.get_latest_harvest_records_by_source_orm(
-            source_id=source_id,
-            count=True,
-        )
-        summary_data = {
-            "records_count": records_count,
-            "last_job_errors": None,
-            "last_job_finished": None,
-            "next_job_scheduled": None,
-            "active_job_in_progress": False,
-        }
-
-        if jobs:
-            last_job = jobs[0]
-            if last_job.status == "in_progress":
-                summary_data["active_job_in_progress"] = True
-
-            last_job_error_count = db.get_harvest_record_errors_by_job(
-                count=True,
-                job_id=last_job.id,
-            )
-            summary_data["last_job_errors"] = last_job_error_count
-            summary_data["last_job_finished"] = last_job.date_finished
-
-        future_jobs = db.get_new_harvest_jobs_by_source_in_future(source_id)
-
-        if future_jobs:
-            summary_data["next_job_scheduled"] = future_jobs[0].date_created
-
-        chart_data_values = dynamic_map_list_items_to_dict(
-            db._to_dict(jobs[::-1]),  # reverse the list order for the chart
-            [
-                "date_finished",
-                "records_added",
-                "records_updated",
-                "records_deleted",
-                "records_errored",
-                "records_ignored",
-            ],
-        )
-        chart_data = {
-            "labels": chart_data_values["date_finished"],
-            "datasets": [
-                {
-                    "label": "Added",
-                    "data": chart_data_values["records_added"],
-                    "borderColor": "green",
-                    "backgroundColor": "green",
-                },
-                {
-                    "label": "Updated",
-                    "data": chart_data_values["records_updated"],
-                    "borderColor": "blue",
-                    "backgroundColor": "blue",
-                },
-                {
-                    "label": "Deleted",
-                    "data": chart_data_values["records_deleted"],
-                    "borderColor": "black",
-                    "backgroundColor": "black",
-                },
-                {
-                    "label": "Errored",
-                    "data": chart_data_values["records_errored"],
-                    "borderColor": "red",
-                    "backgroundColor": "red",
-                },
-                {
-                    "label": "Unchanged",
-                    "data": chart_data_values["records_ignored"],
-                    "borderColor": "grey",
-                    "backgroundColor": "grey",
-                },
-            ],
-        }
-        source = db.get_harvest_source(source_id)
-        datasets_page = request.args.get("datasets_page", 1, type=convert_to_int)
-        datasets_count = db.get_datasets_by_source(source_id=source_id, count=True)
-        datasets_pagination = Pagination(
-            count=datasets_count,
-            current=datasets_page,
-        )
-        datasets = db.get_datasets_by_source(
-            source_id=source_id,
-            page=datasets_pagination.db_current,
-            per_page=datasets_pagination.per_page,
-        )
-        data = {
-            "ckan_url": CKAN_URL,
-            "source": source,
-            "summary_data": summary_data,
-            "jobs": jobs,
-            "chart_data": chart_data,
-            "htmx_vars": jobs_htmx_vars,
-            "datasets": datasets,
-            "datasets_htmx_vars": datasets_htmx_vars,
-        }
-        logger.info(
-            "Rendered harvest source page source_id=%s source_name=%s jobs_total=%s "
-            "jobs_returned=%s datasets_total=%s datasets_returned=%s "
-            "records_count=%s active_job_in_progress=%s",
-            source_id,
-            getattr(source, "name", None),
-            jobs_count,
-            len(jobs),
-            datasets_count,
-            len(datasets),
-            records_count,
-            summary_data["active_job_in_progress"],
-        )
-        return render_template(
-            "view_source_data.html",
-            form=form,
-            pagination=pagination.to_dict(),
-            datasets_pagination=datasets_pagination.to_dict(),
-            data=data,
-        )
-
-
-@main.post("/harvest_source/<source_id>")
-@login_required
-@valid_id_required
-def update_harvest_source_actions(source_id: str):
-    """Handle authenticated harvest source action form submissions."""
-    form = HarvestTriggerForm()
-
-    if not form.validate_on_submit():
-        flash(form.errors)
-        return redirect(url_for("api.view_harvest_source", source_id=source_id))
-
-    if form.edit.data:
-        return redirect(url_for("api.edit_harvest_source", source_id=source_id))
-    elif form.harvest.data:
-        if form.force_check.data:
-            return trigger_manual_job_helper(source_id, "force_harvest")
-        return trigger_manual_job_helper(source_id)
-    elif form.clear.data:
-        return trigger_manual_job_helper(source_id, "clear")
-    elif form.delete.data:
-        try:
-            message, status = db.delete_harvest_source(source_id)
-            _log_mutation("delete", "harvest_source", source_id, status=status)
-            flash(message)
-            if status == 409:
-                return redirect(url_for("api.view_harvest_source", source_id=source_id))
-            else:
-                return redirect(url_for("main.harvest_source_list"))
-        except Exception as e:
-            message = f"Failed to delete harvest source :: {repr(e)}"
-            logger.error(message)
-            flash(message)
-            return redirect(url_for("api.view_harvest_source", source_id=source_id))
-
-    return redirect(url_for("api.view_harvest_source", source_id=source_id))
-
-
-@main.route("/harvest_source_list/", methods=["GET"])
-def harvest_source_list():
-    sources = db.get_all_harvest_sources()
-    data = {"harvest_sources": sources}
-    logger.info("Rendered harvest source list count=%s", len(sources) if sources else 0)
-    return render_template("view_source_list.html", data=data)
-
-
-@main.route("/dataset/<string:dataset_slug>", methods=["GET", "POST"])
-def view_dataset(dataset_slug: str):
-    """View a dataset detail page by slug, and handle slug edits via POST."""
-    dataset = db.get_dataset_by_slug(dataset_slug)
-
-    if request.method == "POST":
-        if not session.get("user"):
-            flash("You must be logged in to edit a dataset slug.")
-            return redirect(url_for("main.view_dataset", dataset_slug=dataset_slug))
-
-        form = DatasetSlugForm(
-            request.form,
-            dataset_id=dataset.id if dataset else None,
-            db_interface=db,
-        )
-        if dataset is None:
-            flash("Dataset not found.")
-            return redirect(url_for("main.harvest_source_list"))
-
-        if form.validate():
-            updated, os_synced, os_error = db.update_dataset_slug(
-                dataset.id, form.slug.data
-            )
-            if updated:
-                _log_mutation(
-                    "edit",
-                    "dataset",
-                    updated.id,
-                    old_slug=dataset_slug,
-                    new_slug=updated.slug,
-                    opensearch_synced=os_synced,
-                )
-                if os_synced:
-                    flash(
-                        f"Slug updated successfully to '{updated.slug}'.",
-                        "success",
-                    )
-                else:
-                    flash(
-                        f"Slug updated to '{updated.slug}' in the database, "
-                        "but the search index could not be updated. "
-                        f"Error: {os_error}",
-                        "warning",
-                    )
-                return redirect(url_for("main.view_dataset", dataset_slug=updated.slug))
-            else:
-                flash(
-                    f"Failed to update slug. Error: {os_error}",
-                    "danger",
-                )
-                return redirect(url_for("main.view_dataset", dataset_slug=dataset_slug))
-        else:
-            # Re-render the page with the form errors shown inline.
-            data = {"dataset": dataset}
-            return (
-                render_template(
-                    "view_dataset_data.html",
-                    data=data,
-                    form=form,
-                ),
-                422,
-            )
-
-    # GET
-    form = DatasetSlugForm(
-        dataset_id=dataset.id if dataset else None,
-        db_interface=db,
-        data={"slug": dataset.slug} if dataset else {},
-    )
-    data = {"dataset": dataset}
-    return render_template(
-        "view_dataset_data.html",
-        data=data,
-        form=form,
-    ), (200 if dataset is not None else 404)
-
-
-### Edit Source (JSON)
-@api.post("/api/harvest_source/edit/<source_id>")
-@api.doc(hide=True)  # don't list the authenticated API
-@login_required
-@valid_id_required
-def edit_harvest_source_api(source_id: str):
-    updated_source = db.update_harvest_source(source_id, request.json)
-    job_message = load_manager.schedule_first_job(updated_source.id)
-
-    if updated_source and job_message:
-        _log_mutation(
-            "edit",
-            "harvest_source",
-            updated_source.id,
-            organization_id=updated_source.organization_id,
-            source_name=updated_source.name,
-        )
-        return {
-            "message": f"Updated source with ID: {updated_source.id}. {job_message}"
-        }, 200
-    else:
-        return {"error": "Failed to update harvest source"}, 400
-
-
-### Edit Source (HTML form)
-@api.route("/harvest_source/edit/<source_id>", methods=["GET", "POST"])
-@api.doc(hide=True)  # don't list the authenticated API
-@login_required
-@valid_id_required  # TODO: Use an HTML 404 page
-def edit_harvest_source(source_id: str):
-    if source_id:
-        source = db.get_harvest_source(source_id)
-        organizations = db.get_all_organizations()
-        if source and organizations:
-            organization_choices = sorted(
-                [
-                    (str(org["id"]), f"{org['name']} - {org['id']}")
-                    for org in db._to_dict(organizations)
-                ],
-                key=lambda x: x[1].lower(),
-            )
-            source_data = db._to_dict(source)
-            source_data["notification_emails"] = ", ".join(
-                source_data.get("notification_emails") or []
-            )
-            form = HarvestSourceForm(data=source_data)
-            form.organization_id.choices = organization_choices
-            if form.validate_on_submit():
-                new_source_data = make_new_source_contract(form)
-                source = db.update_harvest_source(source_id, new_source_data)
-                job_message = load_manager.schedule_first_job(source.id)
-                if source and job_message:
-                    _log_mutation(
-                        "edit",
-                        "harvest_source",
-                        source.id,
-                        organization_id=source.organization_id,
-                        source_name=source.name,
-                    )
-                    flash(f"Updated source with ID: {source.id}. {job_message}")
-                else:
-                    flash("Failed to update harvest source.")
-                return redirect(url_for("api.view_harvest_source", source_id=source.id))
-            elif form.errors:
-                flash(form.errors)
-                return redirect(url_for("api.edit_harvest_source", source_id=source_id))
-            return render_template(
-                "edit_data.html",
-                form=form,
-                action="Edit",
-                data_type="Harvest Source",
-                button="Update",
-                source_id=source_id,
-            )
-        else:
-            flash(f"No source with id: {source_id}")
-            return redirect(url_for("main.harvest_source_list"))
-
-
-# Delete Source
-@api.route("/harvest_source/<source_id>", methods=["DELETE"])
-@api.doc(hide=True)
-@login_required
-@valid_id_required
-def delete_harvest_source(source_id):
-    try:
-        message, status = db.delete_harvest_source(source_id)
-        _log_mutation("delete", "harvest_source", source_id, status=status)
-        return make_response(jsonify({"message": message}), status)
-    except Exception as e:
-        message = f"Failed to delete harvest source :: {repr(e)}"
-        logger.error(message)
-        return make_response(jsonify({"message": message}), 500)
-
-
-### Trigger Harvest
-@api.route("/harvest_source/harvest/<source_id>/<job_type>", methods=["GET"])
-@api.doc(hide=True)  # don't document the authenticated API
-@login_required
-def trigger_harvest_source(source_id, job_type):
-    if not is_valid_uuid4(source_id):
-        return JSON_NOT_FOUND()
-    # possible job_types are in `harvester/harvest.py:harvest_job_starter()`
-    if job_type not in ["harvest", "force_harvest", "clear", "validate"]:
-        return JSON_NOT_FOUND()
-    message = load_manager.trigger_manual_job(source_id, job_type)
-    flash(message)
-    return redirect(url_for("api.view_harvest_source", source_id=source_id))
-
-
-## Harvest Job
-### Add Job (JSON)
-@api.post("/api/harvest_job/add")
-@api.doc(hide=True)  # don't list the autenticated API
-@login_required
-def add_harvest_job():
-    if not request.is_json:
-        return make_response(jsonify({"error": "Request must be JSON"}), 415)
-    job = db.add_harvest_job(request.json)
-    if job:
-        _log_mutation(
-            "create",
-            "harvest_job",
-            job.id,
-            harvest_source_id=job.harvest_source_id,
-            status=job.status,
-        )
-        return {"message": f"Added new harvest job with ID: {job.id}"}, 200
-    else:
-        return {"error": "Failed to add harvest job."}, 400
-
-
-### Get Job (JSON)
-@api.get("/api/harvest_job/<job_id>")
-@api.doc(
-    responses={
-        200: {
-            "description": "View harvest job",
-            "content": {"application/json": {"schema": JobInfo}},
-        }
-    }
-)
-@valid_id_required
-def get_harvest_job(job_id):
-    """Return the JSON representation of a harvest job.
-
-    The HTML detail page lives on the api blueprint at
-    `GET /harvest_job/<job_id>` (see `view_harvest_job`).
-    """
-    job = db.get_harvest_job(job_id)
-    return jsonify(db._to_dict(job)) if job else JSON_NOT_FOUND()
-
-
-### Get Job (HTML detail page)
-@api.route("/harvest_job/<job_id>", methods=["GET"])
-@valid_id_required  # TODO: Use an HTML 404 page
-def view_harvest_job(job_id=None):
-    def _load_json_title(json_string):
-        try:
-            return json.loads(json_string).get("title", None)
-        except Exception as e:
-            logger.error(f"Error loading json source_raw: {repr(e)}")
-            return None
-
-    record_error_count = db.get_harvest_record_errors_by_job(
-        job_id,
-        count=True,
-    )
-    htmx_vars = {
-        "target_div": "#error_results_pagination",
-        "endpoint_url": f"/harvest_job/{job_id}",
-        "page_param": "page",
-    }
-
-    pagination = Pagination(
-        count=record_error_count,
-        current=request.args.get("page", 1, type=convert_to_int),
-    )
-    record_error_summary = db.get_record_errors_summary_by_job(job_id)
-    record_errors = db.get_harvest_record_errors_by_job(
-        job_id, page=pagination.db_current, for_view=True
-    )
-    record_errors_dict = [
-        {
-            "error": make_new_record_error_contract(row),
-            "identifier": row[-2],
-            "title": _load_json_title(row[-1]),  # source_raw
-        }
-        for row in record_errors
-    ]
-
-    if htmx:
-        data = {
-            "harvest_job_id": job_id,
-            "record_errors": record_errors_dict,
-            "htmx_vars": htmx_vars,
-        }
-        logger.info(
-            "Rendered harvest job errors partial job_id=%s page=%s error_count=%s "
-            "returned=%s",
-            job_id,
-            pagination.current,
-            pagination.count,
-            len(record_errors_dict),
-        )
-        return render_block(
-            "view_job_data.html",
-            "record_errors_table",
-            data=data,
-            pagination=pagination.to_dict(),
-        )
-    else:
-        job = db.get_harvest_job(job_id)
-        data = {
-            "job": job,
-            "record_error_summary": record_error_summary,
-            "record_errors": record_errors_dict,
-            "htmx_vars": htmx_vars,
-        }
-        if job and job.status == "in_progress":
-            data["percent_complete"] = process_job_complete_percentage(job.to_dict())
-        logger.info(
-            "Rendered harvest job detail job_id=%s status=%s source_id=%s "
-            "record_error_count=%s record_errors_returned=%s page=%s",
-            job_id,
-            getattr(job, "status", None),
-            getattr(job, "harvest_source_id", None),
-            record_error_count,
-            len(record_errors_dict),
-            pagination.current,
-        )
-
-        return render_template(
-            "view_job_data.html", data=data, pagination=pagination.to_dict()
-        )
-
-
-### Update Job
-@api.route("/harvest_job/<job_id>", methods=["PUT"])
-@api.doc(hide=True)  # don't list the authenticated API
-@login_required
-@valid_id_required
-def update_harvest_job(job_id):
-    result = db.update_harvest_job(job_id, request.json)
-    _log_mutation(
-        "edit",
-        "harvest_job",
-        job_id,
-        harvest_source_id=getattr(result, "harvest_source_id", None),
-        status=getattr(result, "status", None),
-    )
-    return jsonify(db._to_dict(result))
-
-
-### Delete Job
-@api.route("/harvest_job/<job_id>", methods=["DELETE"])
-@api.doc(hide=True)  # don't list the authenticated API
-@login_required
-@valid_id_required
-def delete_harvest_job(job_id):
-    result = db.delete_harvest_job(job_id)
-    _log_mutation("delete", "harvest_job", job_id)
-    return escape(result)
-
-
-@api.route("/harvest_job/cancel/<job_id>", methods=["GET", "POST"])
-@api.doc(hide=True)
-@login_required
-@valid_id_required
-def cancel_harvest_job(job_id):
-    """Cancels a harvest job"""
-    if not is_valid_uuid4(job_id):
-        flash("Invalid job ID.")
-        return redirect("/")
-    message = load_manager.stop_job(job_id)
-    flash(message)
-    return redirect(url_for("api.view_harvest_job", job_id=job_id))
-
-
-### Download all errors for a given job
-@api.route("/harvest_job/<string:job_id>/errors/<string:error_type>", methods=["GET"])
-@api.output(FileSchema, content_type="text/csv")
-def download_harvest_errors_by_job(job_id, error_type):
-    if not is_valid_uuid4(job_id):
-        return JSON_NOT_FOUND()
-    try:
-        if error_type not in ["job", "record"]:
-            return "Invalid error type. Must be 'job' or 'record'", 400
-
-        # Create a temporary file to store the CSV data
-        temp_file = tempfile.NamedTemporaryFile(mode="w+", delete=False, suffix=".csv")
-        temp_file_path = temp_file.name
-
-        try:
-            csv_writer = csv.writer(temp_file)
-
-            if error_type == "job":
-                # Write header
-                header = [
-                    "harvest_job_id",
-                    "date_created",
-                    "job_error_type",
-                    "message",
-                    "harvest_job_error_id",
-                ]
-                csv_writer.writerow(header)
-
-                # Write job errors
-                errors = db._to_dict(db.get_harvest_job_errors_by_job(job_id))
-                for error_dict in errors:
-                    row = [
-                        str(error_dict.get("harvest_job_id", "")),
-                        str(error_dict.get("date_created", "")),
-                        str(error_dict.get("type", "")),
-                        str(error_dict.get("message", "")),
-                        str(error_dict.get("id", "")),
-                    ]
-                    csv_writer.writerow(row)
-
-            elif error_type == "record":
-                # Write header
-                header = [
-                    "record_error_id",
-                    "identifier",
-                    "title",
-                    "harvest_record_id",
-                    "record_error_type",
-                    "message",
-                    "date_created",
-                ]
-                csv_writer.writerow(header)
-
-                # Process record errors in batches
-                page = 0
-                batch_size = 100
-
-                while True:
-                    batch_errors = db.get_harvest_record_errors_by_job(
-                        job_id,
-                        paginate=True,
-                        page=page,
-                        per_page=batch_size,
-                    )
-
-                    if not batch_errors:
-                        break
-
-                    for error, identifier, source_raw in batch_errors:
-                        # Extract title from source_raw JSON
-                        title = ""
-                        if source_raw:
-                            try:
-                                title = json.loads(source_raw).get("title", "")
-                            except (json.JSONDecodeError, AttributeError):
-                                title = ""
-
-                        row = [
-                            str(error.id) if error.id else "",
-                            str(identifier) if identifier else "",
-                            str(title),
-                            (
-                                str(error.harvest_record_id)
-                                if error.harvest_record_id
-                                else ""
-                            ),
-                            str(error.type) if error.type else "",
-                            str(error.message) if error.message else "",
-                            str(error.date_created) if error.date_created else "",
-                        ]
-                        csv_writer.writerow(row)
-
-                    page += 1
-
-                    # If we got fewer results than batch_size, we're done
-                    if len(batch_errors) < batch_size:
-                        break
-
-            temp_file.close()
-
-            # Create a generator that reads the file in chunks
-            def generate_file_chunks():
-                try:
-                    with open(temp_file_path, "r") as f:
-                        while True:
-                            chunk = f.read(8192)  # Read in 8KB chunks
-                            if not chunk:
-                                break
-                            yield chunk
-                finally:
-                    # Clean up the temporary file
-                    try:
-                        os.unlink(temp_file_path)
-                    except OSError:
-                        pass
-
-            # Create streaming response
-            response = Response(
-                generate_file_chunks(),
-                mimetype="text/csv",
-                headers={
-                    "Content-Disposition": f"attachment; filename={job_id}_{error_type}_errors.csv",
-                    "Content-Type": "text/csv",
-                },
-            )
-
-            return response
-
-        except Exception as e:
-            temp_file.close()
-            try:
-                os.unlink(temp_file_path)
-            except OSError:
-                pass
-            raise e
-
-    except Exception as e:
-        logger.error(f"Error in download_harvest_errors_by_job: {repr(e)}")
-        return "Error generating error report", 500
-
-
-# Records
-## Get record
-@api.route("/harvest_record/<record_id>", methods=["GET"])
-@api.output(RecordInfo)
-@valid_id_required
-def get_harvest_record(record_id):
-    record = db.get_harvest_record(record_id)
-    return jsonify(db._to_dict(record)) if record else JSON_NOT_FOUND()
-
-
-## Get records source raw
-@api.route("/harvest_record/<record_id>/raw", methods=["GET"])
-@api.doc(
-    responses={  # nultiple MIME-types, specify them manually
-        200: {
-            "description": "Raw harvest record contents",
-            "content": {
-                "application/json": {"type": "object"},
-                "application/xml": {"type": "object"},
-            },
-        }
-    }
-)
-@valid_id_required
-def get_harvest_record_raw(record_id=None):
-    record = db.get_harvest_record(record_id)
-    if not record or not record.job or not record.job.source:
-        return JSON_NOT_FOUND()
-
-    source = record.job.source
-    schema_type = getattr(source, "schema_type", None)
-
-    # If schema_type contains 'dcatus', treat as JSON, else XML
-    if schema_type and "dcatus" in schema_type:
-        try:
-            source_raw_json = json.loads(record.source_raw)
-            return jsonify(source_raw_json), 200
-        except Exception:
-            logger.error(f"Error returning JSON source raw for record ID: {record_id}")
-            # fallback to raw string if JSON parsing fails
-            return (
-                Response(record.source_raw, mimetype="application/json; charset=utf-8"),
-                200,
-            )
-    else:
-        # Assume XML for all other schema types, including CSDGM and ISO19139
-        # TODO: Consider explicitly handling CSDGM and ISO19139 in case there are other formats in the future
-        return (
-            Response(record.source_raw, mimetype="application/xml; charset=utf-8"),
-            200,
-        )
-
-
-@api.route("/harvest_record/<record_id>/transformed", methods=["GET"])
-@api.doc(
-    responses={
-        200: {
-            "description": "Transformed harvest record contents",
-            "content": {
-                "application/json": {"type": "object"},
-            },
-        }
-    }
-)
-@valid_id_required
-def get_harvest_record_transformed(record_id=None):
-    record = db.get_harvest_record(record_id)
-    if not record:
-        return JSON_NOT_FOUND()
-
-    transformed = record.source_transform
-    if transformed is None:
-        return JSON_NOT_FOUND()
-
-    if isinstance(transformed, str):
-        if not transformed.strip():
-            return JSON_NOT_FOUND()
-        try:
-            transformed = json.loads(transformed)
-        except json.JSONDecodeError:
-            return (
-                Response(transformed, mimetype="application/json; charset=utf-8"),
-                200,
-            )
-
-    return jsonify(transformed), 200
-
-
-## Add record (JSON)
-@api.post("/api/harvest_record/add")
-@api.doc(hide=True)  # don't list the authenticated API
-@login_required
-def add_harvest_record():
-    if not request.is_json:
-        return make_response(jsonify({"error": "Request must be JSON"}), 415)
-    record = db.add_harvest_record(request.json)
-    if record:
-        _log_mutation(
-            "create",
-            "harvest_record",
-            record.id,
-            harvest_job_id=record.harvest_job_id,
-        )
-        return {"message": f"Added new record with ID: {record.id}"}, 200
-    else:
-        return {"error": "Failed to add harvest record."}, 400
-
-
-### Get record errors by record id
-@api.route("/harvest_record/<record_id>/errors", methods=["GET"])
-@api.doc(
-    responses={  # top-level list needs a manual schema
-        200: {
-            "description": "List of errors for this record",
-            "content": {
-                "application/json": {"schema": {"type": "array", "items": ErrorInfo}}
-            },
-        },
-    }
-)
-@valid_id_required
-def get_all_harvest_record_errors(record_id: str) -> list:
-    try:
-        record_errors = db.get_harvest_record_errors_by_record(record_id)
-        return (
-            jsonify(db._to_dict(record_errors)) if record_errors else JSON_NOT_FOUND()
-        )
-    except Exception:
-        return jsonify({"error": "Please provide a valid record_id"}), 404
-
-
-## Harvest Error
-### Get error by id
-@api.route("/harvest_error/<error_id>", methods=["GET"])
-@api.output(ErrorInfo)
-@valid_id_required
-def get_harvest_error(error_id: str = None) -> Response:
-    # retrieves the given error ( either job or record )
-    try:
-        error = db.get_harvest_error(error_id)
-        return jsonify(db._to_dict(error)) if error else JSON_NOT_FOUND()
-    except Exception:
-        return jsonify({"error": "Please provide a valid record_id"}), 404
-
-
-@main.route("/metrics/", methods=["GET"])
-def view_metrics():
-    """Render index page with recent harvest jobs."""
-    current_time = get_datetime()
-    start_time = current_time - timedelta(days=7)
-    time_filter = (
-        f"date_created ge {start_time.isoformat()},date_created le {current_time}"
-    )
-
-    # Handle multiple pagination parameters
-    jobs_page = request.args.get("jobs_page", 1, type=convert_to_int)
-    errors_page = request.args.get("errors_page", 1, type=convert_to_int)
-
-    # Count for pagination
-    count_jobs = db.pget_harvest_jobs(
-        facets=time_filter + ",status eq complete",
-        count=True,
-    )
-
-    count_errors = db.pget_harvest_job_errors(
-        facets=time_filter,
-        count=True,
-    )
-
-    # Create separate pagination objects
-    pagination_jobs = Pagination(
-        count=count_jobs,
-        current=jobs_page,
-    )
-
-    pagination_errors = Pagination(
-        count=count_errors,
-        current=errors_page,
-    )
-
-    # HTMX handling for specific sections
-    if htmx:
-        htmx_target = request.headers.get("HX-Target", "")
-
-        if "paginated__harvest-jobs" in htmx_target:
-            # Handle jobs pagination
-            jobs = db.pget_harvest_jobs(
-                facets=time_filter + ",status eq complete",
-                page=pagination_jobs.db_current,
-                per_page=pagination_jobs.per_page,
-                order_by="desc",
-            )
-            htmx_vars = {
-                "target_div": "#paginated__harvest-jobs",
-                "endpoint_url": "/metrics/",
-                "page_param": "jobs_page",
-            }
-            data = {
-                "jobs": jobs,
-                "htmx_vars": htmx_vars,
-            }
-            logger.info(
-                "Rendered metrics jobs partial page=%s total_jobs=%s returned=%s "
-                "window_start=%s window_end=%s",
-                pagination_jobs.current,
-                count_jobs,
-                len(jobs),
-                start_time.isoformat(),
-                current_time.isoformat(),
-            )
-            return render_block(
-                "metrics_dashboard.html",
-                "htmx_paginated_jobs",
-                data=data,
-                pagination_jobs=pagination_jobs.to_dict(),
-            )
-
-        elif "paginated__harvest-errors" in htmx_target:
-            # Handle errors pagination
-            errors_time_filter = f"date_created ge {start_time.isoformat()},date_created le {current_time}"
-            failures = db.pget_harvest_job_errors(
-                facets=errors_time_filter,
-                page=pagination_errors.db_current,
-                per_page=pagination_errors.per_page,
-                order_by="desc",
-            )
-            htmx_vars = {
-                "target_div": "#paginated__harvest-errors",
-                "endpoint_url": "/metrics/",
-                "page_param": "errors_page",
-            }
-            data = {
-                "failures": failures,
-                "htmx_vars": htmx_vars,
-            }
-            logger.info(
-                "Rendered metrics errors partial page=%s total_errors=%s returned=%s "
-                "window_start=%s window_end=%s",
-                pagination_errors.current,
-                count_errors,
-                len(failures),
-                start_time.isoformat(),
-                current_time.isoformat(),
-            )
-            return render_block(
-                "metrics_dashboard.html",
-                "htmx_paginated_errors",
-                data=data,
-                pagination_errors=pagination_errors.to_dict(),
-            )
-    else:
-        # Full page load
-        jobs = db.pget_harvest_jobs(
-            facets=time_filter + ",status eq complete",
-            page=pagination_jobs.db_current,
-            per_page=pagination_jobs.per_page,
-            order_by="desc",
-        )
-        errors_time_filter = (
-            f"date_created ge {start_time.isoformat()},date_created le {current_time}"
-        )
-        failures = db.pget_harvest_job_errors(
-            facets=errors_time_filter,
-            page=pagination_errors.db_current,
-            per_page=pagination_errors.per_page,
-            order_by="desc",
-        )
-        current_jobs = db.pget_harvest_jobs(
-            facets="status eq in_progress",
-            order_by="desc",
-        )
-
-        for job in current_jobs:
-            job.percent_complete = process_job_complete_percentage(job.to_dict())
-
-        data = {
-            "current_jobs": current_jobs,
-            "jobs": jobs,
-            "new_jobs_in_past": db.get_new_harvest_jobs_in_past(),
-            "failures": failures,
-            "current_time": current_time,
-            "window_start": start_time,
-        }
-        logger.info(
-            "Rendered metrics page current_jobs=%s completed_jobs_total=%s "
-            "completed_jobs_returned=%s failure_total=%s failure_returned=%s "
-            "window_start=%s window_end=%s",
-            len(current_jobs),
-            count_jobs,
-            len(jobs),
-            count_errors,
-            len(failures),
-            start_time.isoformat(),
-            current_time.isoformat(),
-        )
-        return render_template(
-            "metrics_dashboard.html",
-            pagination_jobs=pagination_jobs.to_dict(),
-            pagination_errors=pagination_errors.to_dict(),
-            data=data,
-        )
-
-
-# Builder Query for JSON feed
-@api.route("/organizations/", methods=["GET"])
-@api.route("/harvest_sources/", methods=["GET"])
-@api.route("/harvest_records/", methods=["GET"])
-@api.route("/harvest_jobs/", methods=["GET"])
-@api.route("/harvest_job_errors/", methods=["GET"])
-@api.route("/harvest_record_errors/", methods=["GET"])
-@api.input(
-    QueryInfo, location="query", validation=False
-)  # we handle the request directly
-@api.doc(
-    responses={  # list return type requires a manual schema
-        200: {
-            "description": "List of reuslts for this query",
-            "content": {
-                "application/json": {
-                    "schema": {
-                        "type": "array",
-                        "items": {
-                            "oneOf": [
-                                ErrorInfo,
-                                OrgInfo,
-                                SourceInfo,
-                                RecordInfo,
-                                JobInfo,
-                            ]
-                        },
-                    }
-                },
-            },
-        }
-    }
-)
-def json_builder_query(**kwargs):
-    job_id = request.args.get("harvest_job_id")
-    source_id = request.args.get("harvest_source_id")
-    facets = request.args.get("facets", default="")
-
-    if job_id is not None:
-        if facets:
-            facets += f",harvest_job_id eq {job_id}"
-        else:
-            facets = f"harvest_job_id eq {job_id}"
-    if source_id is not None:
-        if facets:
-            facets += f",harvest_source_id eq {source_id}"
-        else:
-            facets = f"harvest_source_id eq {source_id}"
-
-    model = escape(request.path).replace("/", "")
-    try:
-        res = db.pget_db_query(
-            model=model,
-            page=request.args.get("page", type=convert_to_int),
-            per_page=request.args.get("per_page", type=convert_to_int),
-            paginate=request.args.get("paginate", type=is_it_true),
-            count=request.args.get("count", type=is_it_true),
-            order_by=request.args.get("order_by"),
-            facets=facets,
-        )
-        if not res:
-            return f"No {model} found for this query", 404
-        elif isinstance(res, int):
-            # in case we are just returning a count from db query
-            return {"count": res, "type": model}
-        else:
-            return jsonify(db._to_dict(res))
-    except Exception as e:
-        logger.info(f"Failed json_builder_query :: {repr(e)} ")
-        return "Error with query", 400
-
-
-@main.route("/openapi/docs")
-def openapi_docs():
-    return render_template("/swagger.html")
-
-
-@api.route("/api/validate", methods=["POST"])
-@api.input(ValidatorInfo)
-@api.output(ValidationResultSchema, status_code=200)
-@api.doc(
-    summary="Validate a DCAT catalog against a v1.1 or v3.0 schema",
-    description="Downloads or parses a DCATUS catalog and validates each dataset.",
-    responses={
-        500: {
-            "description": "Failed to download or process the catalog",
-            "content": {"application/json": {"schema": ValidationErrorResponseSchema}},
-        },
-    },
-)
-def validator(json_data):
-    """
-    api route for validating v1.1 or v3.0 dcatus catalogs
-    """
-    errors = []
-    data = []
-
-    try:
-        if json_data["fetch_method"] == "url":
-            data = fetch_json_from_url(json_data["url"])
-
-        if json_data["fetch_method"] == "paste":
-            data = json.loads(json_data["json_text"])
-
-        errors = validate_records(data, json_data["schema"])
-        logger.info(
-            "API validator completed fetch_method=%s schema=%s validation_errors=%s",
-            json_data["fetch_method"],
-            json_data["schema"],
-            len(errors),
-        )
-    except Exception as e:
-        logger.error(f"API Validator error :: {repr(e)}")
-        return make_response(
-            jsonify(
-                {"error": "API Validator error: failed to validate dcatus catalog"}
-            ),
-            500,
-        )
-
-    return make_response(
-        jsonify({"validation_errors": errors}),
-        200,
-    )
-
-
-@main.route("/validate/", methods=["GET", "POST"])
-def view_validators():
-    """
-    view for validating v1.1 or v3.0 dcatus catalogs using form
-    """
-    data = []
-    errors = []
-    submitted = False
-
-    form = ValidatorForm()
-    if form.validate_on_submit():
-        if form.fetch_method.data == "url":
-            try:
-                data = fetch_json_from_url(form.url.data)
-            except Exception as e:
-                form.url.errors.append(str(e))
-        elif form.fetch_method.data == "paste":
-            try:
-                data = json.loads(form.json_text.data)
-            except Exception as e:
-                form.json_text.errors.append(str(e))
-        elif form.fetch_method.data == "upload":
-            try:
-                raw = form.json_file.data.read()
-                data = json.loads(raw)
-            except json.JSONDecodeError as e:
-                form.json_file.errors.append(f"Invalid JSON in uploaded file: {e}")
-            except Exception as e:
-                form.json_file.errors.append(str(e))
-
-        if not form.errors:
-            submitted = True
-            errors = validate_records(data, form.schema.data)
-            logger.info(
-                "Rendered validator results fetch_method=%s schema=%s validation_errors=%s",
-                form.fetch_method.data,
-                form.schema.data,
-                len(errors),
-            )
-        else:
-            logger.warning(
-                "Validator submission failed fetch_method=%s url_errors=%s json_errors=%s",
-                form.fetch_method.data,
-                len(form.url.errors),
-                len(form.json_text.errors),
-            )
-
-    data = {
-        "record_errors": errors,
-    }
-
-    if request.method == "GET":
-        logger.info("Rendered validator page")
-
-    return render_template(
-        "view_validators.html",
-        form=form,
-        data=data,
-        submitted=submitted,
-    )
-
-
-def register_routes(app):
-    app.register_blueprint(main)
-    app.register_blueprint(api)
+from app.blueprints.main.auth import create_client_assertion
+
+__all__ = [
+    "UnsafeTemplateEnvError",
+    "create_client_assertion",
+    "db",
+    "load_manager",
+    "register_routes",
+    "render_block",
+]
