+++ app/templates/view_dataset_data.html
@@ -28,7 +28,7 @@
             <tr>
                 <td>Harvest Source:</td>
                 <td>
-                    <a href="{{ url_for('api.view_harvest_source', source_id=data.dataset.harvest_source_id) }}">
+                    <a href="{{ url_for('main.view_harvest_source', source_id=data.dataset.harvest_source_id) }}">
                         {{ data.dataset.harvest_source_id }}
                     </a>
                 </td>
