+++ app/templates/view_job_data.html
@@ -34,7 +34,7 @@
             <tr>
                 {% if key == 'harvest_source_id' %}
                 <td>Harvest Source:</td>
-                <td><a href="{{ url_for('api.view_harvest_source', source_id=value) }}">{{data.job.source.name}}</a></td>
+                <td><a href="{{ url_for('main.view_harvest_source', source_id=value) }}">{{data.job.source.name}}</a></td>
                 {% elif key == 'date_created' or key == 'date_finished' %}
                 <td>{{key}}:</td>
                 <td class="utc-date" data-utc-date="{{ value | utc_isoformat }}">
